refactor: add slottable datepicker fields
